requirements.txt
Flask==1.1.1
python-dotenv==0.10.3
google-hangouts-chat-bot==0.1.4
google-cloud-datastore==1.10.0
